add cart link to the navbar and rename cart and cartitem component to jsx
